All tests pass for minimax algorithm
I had quite a few bugs. Most notable were issues with counting
negative scores as a good thing when predicting the moves
of an opponent.

The tests include a variety of one-off moves that show the ai is
capable of making the right choice offensively and defensively.
A test with two AI playing against each other,
which is expected to end in a draw, is also included.